Display errors when running nosetests
During Ubuntu builds we run the run_test.sh script
during the package build process, so that we can catch
regressions in our packaging before uploading to the archive.

However a recent change writes the output of the nosetests
to an error log so we don't see why the package fail to build.

This patch also displays the output of the nosetests so we
can see why the tests fail without having to look at the nosetests.log.

This patch add the '-v', for a verbose display of nosetests.

Change-Id: Id21d53df356de344a8577282c5a82f3ff927bc11
Signed-off-by: Chuck Short <chuck.short@canonical.com>