Fix rpc control_exchange regression.
A change was made in rpc that requires project's to define the
control_exchange option themselves.  This was so the project's could
set their own default value.  This change has made it in to nova without
adding the control_exchange option, meaning that it is using the default
of 'openstack', potentially conflicting with other projects.

See https://review.openstack.org/#/c/12518/

Fix bug 1083944.

Change-Id: Ifc00a537c4ee327cbee91fe1d911f8c402e350b7