Reset volume_api before cinder cloud tests.
If tests are not run in the default order nosetests runs them in it is
possible that volume_api in test_cinder_cloud.py is in a state
unsuitable for testing. Reset the volume_api before each test to ensure
volume_api is usable for every test run regardless of test order.

Part of blueprint grizzly-testtools

Change-Id: I974e9eb09cbf9f19a01ecd0f13b2664291a5609e